Move liveEditing features into a separate file
